+++ src/c_two/__init__.py
@@ -10,3 +10,31 @@ from .crm.meta import icrm, iicrm
 from .rpc.transferable import transfer, auto_transfer, Transferable, transferable
 
 __version__ = '0.2.2'
+
+LOGO_ASCII ="""
+                                                  
+                                                  
+                       cccc                       
+                   cccccccccccc                   
+               ccc  ccccccccc    cc               
+          cccc      ccccccccc        ccc          
+      ccccccc       cccc cccc        ccccccc      
+     cccccccc       cccc cccc        ccccccc      
+     cccccccc       cccc cccc        ccccccc      
+     ccccccc        cccc cccc        ccccccc      
+     ccc            cccc cccc        ccc          
+                    cccc cccc                     
+                    cccc cccc                     
+                    cccc cccc                     
+           cc       cccc cccc              c      
+       cccccc       cccc cccc         cccccc      
+     cccccccc       cccc cccc        ccccccc      
+     cccccccc       cccc cccc        ccccccc      
+     cccccccc       cccc cccc        ccccccc      
+         cccc       cccc cccc       cccc          
+              cc    cccc cccc   cccc              
+                  cccccc ccccccc                  
+                      cc cc                       
+                                                  
+                                                  
+"""
