+++ src/c_two/cli.py
@@ -1,18 +1,15 @@
-from . import __version__
-
-# C-Two Command-Line Interface (cccli)
 import click
 import logging
+from .seed import build
+from . import __version__, LOGO_ASCII
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
-from .seed import build
-
 @click.group()
 @click.version_option(version=__version__, prog_name='C-Two')
 def cli():
-    """C-Two Command-Line Interface (c2)"""
-    pass
+    """C-Two Command-Line Interface (cccli, c3)"""
 
 cli.command(build)
+print(LOGO_ASCII)
